==1.6== Added: Taketicket is now usable from the console to make it possible for ticket management via website. Command: /taketicket <id> <assignee>
Bugfixes:
Some missing code made MaxTicket when using MySQL unusable.

Progressing:
Added a new table sht_hus, for our custom house inspections on Maera.
A command for this is also in progress and will require the player to be
in a chunk that him/her owns.
Same idea for people looking to advance to Proville is also in progress.

Warning: